+++ lib/src/Formatters/ExtensionDeclarationFormatter.dart
@@ -1,5 +1,6 @@
 import 'package:analyzer/dart/ast/ast.dart';
 
+import '../Data/ConfigExtension.dart';
 import 'TypedFormatter.dart';
 
 class ExtensionDeclarationFormatter extends TypedFormatter<ExtensionDeclaration>
@@ -10,10 +11,10 @@ class ExtensionDeclarationFormatter extends TypedFormatter<ExtensionDeclaration>
     void formatNode(ExtensionDeclaration node)
     {
         formatState.acceptList(node.sortedCommentAndAnnotations, astVisitor, '$methodName/node.sortedCommentAndAnnotations');
-        formatState.copyEntity(node.augmentKeyword, astVisitor, '$methodName/node.augmentKeyword');
-        formatState.copyEntity(node.extensionKeyword, astVisitor, '$methodName/node.extensionKeyword');
-        formatState.copyEntity(node.typeKeyword, astVisitor, '$methodName/node.typeKeyword');
-        formatState.copyEntity(node.name, astVisitor, '$methodName/node.name');
+        copyZeroOne(node, node.augmentKeyword, '$methodName/node.augmentKeyword');
+        copyZeroOne(node, node.extensionKeyword, '$methodName/node.extensionKeyword');
+        copyZeroOne(node, node.typeKeyword, '$methodName/node.typeKeyword');
+        copyZeroOne(node, node.name, '$methodName/node.name');
         formatState.copyEntity(node.typeParameters, astVisitor, '$methodName/node.typeParameters');
         formatState.copyEntity(node.onClause, astVisitor, '$methodName/node.onClause');