+++ lib/src/Formatters/FieldDeclarationFormatter.dart
@@ -11,12 +11,12 @@ class FieldDeclarationFormatter extends TypedFormatter<FieldDeclaration>
     void formatNode(FieldDeclaration node)
     {
         formatState.acceptList(node.metadata, astVisitor, '$methodName/node.metadata');
-        formatState.copyEntity(node.abstractKeyword, astVisitor, '$methodName/node.abstractKeyword');
-        formatState.copyEntity(node.augmentKeyword, astVisitor, '$methodName/node.augmentKeyword');
-        formatState.copyEntity(node.covariantKeyword, astVisitor, '$methodName/node.covariantKeyword');
-        formatState.copyEntity(node.externalKeyword, astVisitor, '$methodName/node.externalKeyword');
-        formatState.copyEntity(node.staticKeyword, astVisitor, '$methodName/node.staticKeyword');
-        formatState.copyEntity(node.fields, astVisitor, '$methodName/node.fields', config.getSpacesZeroOne(node, node.fields));
+        copyZeroOne(node, node.abstractKeyword, '$methodName/node.abstractKeyword');
+        copyZeroOne(node, node.augmentKeyword, '$methodName/node.augmentKeyword');
+        copyZeroOne(node, node.covariantKeyword, '$methodName/node.covariantKeyword');
+        copyZeroOne(node, node.externalKeyword, '$methodName/node.externalKeyword');
+        copyZeroOne(node, node.staticKeyword, '$methodName/node.staticKeyword');
+        copyZeroOne(node, node.fields, '$methodName/node.fields');
         formatState.copySemicolon(node.semicolon, config, '$methodName/node.semicolon', config.space0);
     }
 }