+++ lib/src/Formatters/MixinDeclarationFormatter.dart
@@ -1,5 +1,6 @@
 import 'package:analyzer/dart/ast/ast.dart';
 
+import '../Data/ConfigExtension.dart';
 import 'TypedFormatter.dart';
 
 class MixinDeclarationFormatter extends TypedFormatter<MixinDeclaration>
@@ -10,10 +11,10 @@ class MixinDeclarationFormatter extends TypedFormatter<MixinDeclaration>
     void formatNode(MixinDeclaration node)
     {
         formatState.acceptList(node.sortedCommentAndAnnotations, astVisitor, '$methodName/node.sortedCommentAndAnnotations');
-        formatState.copyEntity(node.augmentKeyword, astVisitor, '$methodName/node.augmentKeyword');
-        formatState.copyEntity(node.baseKeyword, astVisitor, '$methodName/node.baseKeyword');
-        formatState.copyEntity(node.mixinKeyword, astVisitor, '$methodName/node.mixinKeyword');
-        formatState.copyEntity(node.name, astVisitor, '$methodName/node.name');
+        copyZeroOne(node, node.augmentKeyword, '$methodName/node.augmentKeyword');
+        copyZeroOne(node, node.baseKeyword, '$methodName/node.baseKeyword');
+        copyZeroOne(node, node.mixinKeyword, '$methodName/node.mixinKeyword');
+        copyZeroOne(node, node.name, '$methodName/node.name');
         formatState.copyEntity(node.typeParameters, astVisitor, '$methodName/node.typeParameters');
         formatState.copyEntity(node.onClause, astVisitor, '$methodName/node.onClause');
         formatState.copyEntity(node.implementsClause, astVisitor, '$methodName/node.implementsClause');