+++ lib/src/Formatters/TopLevelVariableDeclarationFormatter.dart
@@ -11,9 +11,9 @@ class TopLevelVariableDeclarationFormatter extends TypedFormatter<TopLevelVariab
     void formatNode(TopLevelVariableDeclaration node)
     {
         formatState.acceptList(node.sortedCommentAndAnnotations, astVisitor, '$methodName/node.sortedCommentAndAnnotations');
-        formatState.copyEntity(node.augmentKeyword, astVisitor, '$methodName/node.augmentKeyword');
-        formatState.copyEntity(node.externalKeyword, astVisitor, '$methodName/node.externalKeyword');
-        formatState.copyEntity(node.variables, astVisitor, '$methodName/node.variables');
+        copyZeroOne(node, node.augmentKeyword, '$methodName/node.augmentKeyword');
+        copyZeroOne(node, node.externalKeyword, '$methodName/node.externalKeyword');
+        copyZeroOne(node, node.variables, '$methodName/node.variables');
         formatState.copySemicolon(node.semicolon, config, '$methodName/node.semicolon', config.space0);
     }
 }