+++ test/Formatters/AssertInitializerFormatter_test.dart
@@ -44,21 +44,6 @@ void main()
         TestGroupConfig(
             inputNodeCreator: AstCreator.createConstructorInitializer,
             inputLeading: 'class C{C():',
-            inputMiddle: "assert(true,'message')",
-            inputTrailing: ';}',
-            name: "AssertInitializer assert(true, 'message')",
-            astVisitors: <TestVisitor<void>>[
-                TestVisitor<BooleanLiteral>(19, 'true'),
-                TestVisitor<SimpleStringLiteral>(24, "'message'")
-            ],
-            testConfigs: <TestConfig>[
-                TestConfig.none(),
-                TestConfig("assert(true,'message')")
-            ]
-        ),
-        TestGroupConfig(
-            inputNodeCreator: AstCreator.createConstructorInitializer,
-            inputLeading: 'class C{C():',
             inputMiddle: "assert(true,'message',)",
             inputTrailing: ';}',
             name: "AssertInitializer assert(true, 'message',) (too little spacing)",
@@ -70,8 +55,8 @@ void main()
                 TestConfig.none(),
                 TestConfig.custom('Custom1', Config.all(fixSpaces: false), "assert(true,'message')"),
                 // ignore: avoid_redundant_argument_values
-                TestConfig.custom('Custom2', Config.all(fixSpaces: true, removeTrailingCommas: false), "assert(true,'message',)"),
-                TestConfig("assert(true,'message')")
+                TestConfig.custom('Custom2', Config.all(fixSpaces: true, removeTrailingCommas: false), "assert(true, 'message',)"),
+                TestConfig("assert(true, 'message')")
             ]
         ),
         TestGroupConfig(
@@ -79,17 +64,17 @@ void main()
             inputLeading: 'class C{C():',
             inputMiddle: "assert  (  true  ,  'message'  ,  )",
             inputTrailing: ';}',
-            name: "AssertInitializer assert(true, 'message',) (too much spacing)", // TODO: too little spacing
+            name: "AssertInitializer assert(true, 'message',)",
             astVisitors: <TestVisitor<void>>[
-                TestVisitor<BooleanLiteral>(21, '  true'),
-                TestVisitor<SimpleStringLiteral>(30, "  'message'")
+                TestVisitor<BooleanLiteral>(23, 'true'),
+                TestVisitor<SimpleStringLiteral>(32, "'message'")
             ],
             testConfigs: <TestConfig>[
                 TestConfig.none(),
                 TestConfig.custom('Custom1', Config.all(fixSpaces: false), "assert  (  true  ,  'message'    )"),
                 // ignore: avoid_redundant_argument_values
-                TestConfig.custom('Custom2', Config.all(fixSpaces: true, removeTrailingCommas: false), "assert(  true,  'message',)"),
-                TestConfig("assert(  true,  'message')")
+                TestConfig.custom('Custom2', Config.all(fixSpaces: true, removeTrailingCommas: false), "assert(true, 'message',)"),
+                TestConfig("assert(true, 'message')")
             ]
         )
     ];